requirements.txt
youtube-transcript-api
nltk
scikit-learn
transformers
torch
openai-whisper
torchaudio
pytube
streamlit
